Fix merge conflicts and libc dependency
This updates the libc dependency to use the patch mechanism at the
workspace level. Previously there were two versions of libc in use, and
now there is just one.

The patch must happen in the std Cargo.toml and not just the workspace
level Cargo.toml because of this build-std limitation:
https://github.com/rust-lang/wg-cargo-std-aware/issues/78#issuecomment-1013894681