Support std threads again by using pthread
Priority and affinity are passed through special pthread attr functions
which will be added to our libc fork.

The current thread's priority also needs to be fetched since it is the
default priority value. This is done using another new pthread function,
though this value/mechanism isn't exposed as part of std's API.